Add printing again

+++ src/dimensions.rs
@@ -1,6 +1,6 @@
 use crate::point::{Point, PointTy};
 
-pub type DimTy = u8;
+pub type DimTy = usize;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct Dimensions {
