+++ src/point.rs
@@ -9,7 +9,7 @@ use rustsat::types::Lit;
 use crate::grid::Grid;
 // The math is a mess but whatever
 
-pub type PointTy = i16;
+pub type PointTy = isize;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Default)]
 pub struct Point {
